Add arabic texts

+++ lib/features/orders/models/driver_model.dart
@@ -39,7 +39,7 @@ class Drivers {
   String? fcmToken;
   String? createdAt;
   String? updatedAt;
-  int? distance;
+  double? distance;
   Driver? driver;
 
   Drivers(
